Refactoring JWT Token genration
